+++ packages/db/src/search/GranuleSearch.ts
@@ -53,21 +53,30 @@ export class GranuleSearch extends BaseSearch {
       collections: collectionsTable,
       providers: providersTable,
       pdrs: pdrsTable,
-      granuleGroups: granuleGroupsTable
+      granuleGroups: granuleGroupsTable,
     } = TableNames;
     const countQuery = knex(this.tableName)
       .count('*');
 
-    const searchQuery = knex(this.tableName)
-      .select(`${this.tableName}.*`)
-      .select({
-        providerName: `${providersTable}.name`,
-        collectionName: `${collectionsTable}.name`,
-        collectionVersion: `${collectionsTable}.version`,
-        pdrName: `${pdrsTable}.name`,
+    const searchParams = {
+      providerName: `${providersTable}.name`,
+      collectionName: `${collectionsTable}.name`,
+      collectionVersion: `${collectionsTable}.version`,
+      pdrName: `${pdrsTable}.name`,
+    };
+
+    const fullSearchParams = this.dbQueryParameters.includeActiveState ?
+      {
         groupId: `${granuleGroupsTable}.group_id`,
         state: `${granuleGroupsTable}.state`,
-      })
+        ...searchParams,
+      } : searchParams;
+
+    const searchQuery = knex(this.tableName)
+      .select(`${this.tableName}.*`)
+      .select(
+        fullSearchParams
+      )
       .innerJoin(collectionsTable, `${this.tableName}.collection_cumulus_id`, `${collectionsTable}.cumulus_id`);
 
     if (this.searchCollection()) {
@@ -88,11 +97,10 @@ export class GranuleSearch extends BaseSearch {
       searchQuery.leftJoin(pdrsTable, `${this.tableName}.pdr_cumulus_id`, `${pdrsTable}.cumulus_id`);
     }
 
-    if(this.dbQueryParameters.includeActiveStatus){
-      searchQuery.innerJoin(granuleGroupsTable, `${this.tableName}.cumulus_id`, `${granuleGroupsTable}.granule_cumulus_id`)
-    } else {
-      searchQuery.leftJoin(granuleGroupsTable, `${this.tableName}.cumulus_id`, `${granuleGroupsTable}.granule_cumulus_id`)
+    if (this.dbQueryParameters.includeActiveState) {
+      searchQuery.innerJoin(granuleGroupsTable, `${this.tableName}.cumulus_id`, `${granuleGroupsTable}.granule_cumulus_id`);
     }
+
     return { countQuery, searchQuery };
   }
 
@@ -192,7 +200,8 @@ export class GranuleSearch extends BaseSearch {
       const pdr = item.pdrName ? { name: item.pdrName } : undefined;
       const providerPgRecord = item.providerName ? { name: item.providerName } : undefined;
       const fileRecords = fileMapping[granulePgRecord.cumulus_id] || [];
-      const granuleGroupRecord = item.state && item.groupId ? { state: item.state, group_id: item.groupId} : undefined;
+      const granuleGroupRecord = item.state && item.groupId ?
+        { state: item.state, group_id: item.groupId } : undefined;
       const apiRecord = translatePostgresGranuleToApiGranuleWithoutDbQuery({
         granulePgRecord,
         collectionPgRecord,